tail split across modules
