+++ jhack/charm/record.py
@@ -8,7 +8,8 @@ import typer
 from jhack.helpers import get_current_model
 from jhack.logger import logger
 from jhack.utils.show_relation import RelationData, get_relation_data, get_relations
-from jhack.utils.tail_charms import EventLogMsg, _tail_events
+from jhack.utils.tail_charms.tail_charms import tail_charms
+from jhack.utils.tail_charms.core.parser import EventLogMsg
 
 KeyValueMapping = Dict[str, str]
 
@@ -41,7 +42,9 @@ class Recorder:
         self._output = output
 
     def record(self):
-        _tail_events([self._unit], replay=False, add_new_targets=False, _on_event=self._on_event)
+        tail_charms(
+            [self._unit], replay=False, add_new_targets=False, _on_event=self._on_event
+        )
         self._dump_json()
         return self._state_history
 
@@ -61,7 +64,10 @@ class Recorder:
         app = self._app
         for relation in get_relations(model):
             logger.debug(f"found relation {relation}")
-            if relation.requirer.split(":")[0] == app or relation.provider.split(":")[0] == app:
+            if (
+                relation.requirer.split(":")[0] == app
+                or relation.provider.split(":")[0] == app
+            ):
                 relation_data = get_relation_data(
                     provider_endpoint=relation.provider,
                     requirer_endpoint=relation.requirer,
