+++ jhack/utils/tail_charms/core/deferral_status.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class DeferralStatus(str, enum.Enum):
+    null = "null"
+    deferred = "deferred"
+    reemitted = "reemitted"
+    bounced = "bounced"
